requirements.txt
streamlit
google-genai
google-generativeai
python-dotenv
langchain
Pypdf2
chromadb